Introduce dedicated resizing component
