+++ src/components/Grid.tsx
@@ -1,4 +1,4 @@
-import { type FC, type ReactEventHandler, useCallback, useEffect, useRef } from 'react';
+import { forwardRef, type ReactEventHandler, useCallback } from 'react';
 import { type Board, getBoardColor, getBoardHeight, getBoardWidth } from '../board';
 import type { Color } from '../color';
 import { CirclePath } from './CirclePath';
@@ -12,10 +12,8 @@ export interface GridProps {
 const SIZE = 35;
 const OVERLAP = 0;
 
-export const Grid: FC<GridProps> = ({ board, onClick }) => {
-  const parentRef = useRef<HTMLDivElement>(null);
-  const svgRef = useRef<SVGSVGElement>(null);
-
+export const Grid = forwardRef<SVGSVGElement, GridProps>((props, ref) => {
+  const { board, onClick } = props;
   const size = SIZE;
   const width = getBoardWidth(board);
   const height = getBoardHeight(board);
@@ -29,70 +27,44 @@ export const Grid: FC<GridProps> = ({ board, onClick }) => {
     onClick(getBoardColor(board, x, y)!);
   }, [board, onClick]);
 
-  useEffect(() => {
-    const parent = parentRef.current;
-    const svg = svgRef.current;
-    if (!parent || !svg) return;
-    const callback = (entries: ResizeObserverEntry[]) => {
-      const originalWidth = entries[0].contentRect.width;
-      const flooredWidth = Math.floor(originalWidth / width) * width;
-      svg.style.width = flooredWidth / originalWidth * 100 + '%';
-    };
-    const observer = new ResizeObserver(callback);
-    observer.observe(parent);
-    return () => {
-      observer.disconnect();
-    };
-  }, []);
-
   return (
-    <div
-      ref={parentRef}
+    <svg
+      ref={ref}
+      className="Grid"
+      viewBox={VIEWBOX}
+      preserveAspectRatio="none"
       style={{
-        width: '100%',
-        height: '100%',
-        display: 'grid',
-        placeItems: 'center',
+        borderRadius: '4px',
+        boxShadow: 'var(--box-shadow)',
       }}
     >
-      <svg
-        ref={svgRef}
-        className="Grid"
-        viewBox={VIEWBOX}
-        preserveAspectRatio="none"
-        style={{
-          borderRadius: '4px',
-          boxShadow: 'var(--box-shadow)',
-        }}
-      >
-        {Array.from({ length: height }, (_, y) => [
-          Array.from({ length: width }, (_, x) => (
-            <GridSquare
-              key={x + ',' + y}
-              x={x * size - OVERLAP}
-              y={y * size - OVERLAP}
-              width={size + OVERLAP * 2}
-              height={size + OVERLAP * 2}
-              onPointerDown={handleClick}
-              data-x={x}
-              data-y={y}
-              distance={Math.sqrt(x ** 2 + y ** 2)}
-              color={getBoardColor(board, x, y)!}
-            />
-          ))
-        ])}
+      {Array.from({ length: height }, (_, y) => [
+        Array.from({ length: width }, (_, x) => (
+          <GridSquare
+            key={x + ',' + y}
+            x={x * size - OVERLAP}
+            y={y * size - OVERLAP}
+            width={size + OVERLAP * 2}
+            height={size + OVERLAP * 2}
+            onPointerDown={handleClick}
+            data-x={x}
+            data-y={y}
+            distance={Math.sqrt(x ** 2 + y ** 2)}
+            color={getBoardColor(board, x, y)!}
+          />
+        ))
+      ])}
 
-        <CirclePath
-          outerRadius={size * 0.3}
-          innerRadius={size * 0.2}
-          cx={size / 2}
-          cy={size / 2}
-          fill="#fff"
-          style={{
-            pointerEvents: 'none',
-          }}
-        />
-      </svg>
-    </div>
+      <CirclePath
+        outerRadius={size * 0.3}
+        innerRadius={size * 0.2}
+        cx={size / 2}
+        cy={size / 2}
+        fill="#fff"
+        style={{
+          pointerEvents: 'none',
+        }}
+      />
+    </svg>
   );
-};
+});
